split env
